# target/classes/open-api/api.yaml
openapi: 3.0.3
info:
  title: Coding Challenge Service
  description: Coding Challenge Service
  version: 1.0.0
servers:
  - url: https://app.swaggerhub.com/apis/organizationxx/coding-challenge_service/1.0.0
tags:
  - name: coding-challenge-service
    description: Operations related to players and gaming activity
paths:
  /players/largest-ggr:
    get:
      tags:
        - coding-challenge-controller
      summary: Retrieve player with largest GGR
      description: Retrieve player with largest GGR, matching the pagination filter.
      operationId: get-player-with-largest-GGR
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: pageSize
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/PlayerGGR'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        '404':
          description: Players not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'

components:
  schemas:
    PlayerGGR:
      type: object
      properties:
        id:
          type: integer
          format: int64
        ggr:
          type: number

    ErrorDetails:
      title: ErrorDetails
      type: object
      properties:
        type:
          type: string
          description: A URI reference that identifies the problem type.
          nullable: true
        title:
          type: string
          description: 'Short, human-readable summary of the error type.'
          nullable: true
        status:
          type: integer
          description: The HTTP status code
          format: int32
          nullable: true
        detail:
          type: string
          description: Human-readable explanation specific to this occurrence of the problem
          nullable: true